+++ srcs/ConfigConstant.cpp
@@ -11,7 +11,7 @@ std::string default_index[] = {"index.html"};
 const std::vector<std::string>
     ConfigConstant::DEFAULT_ALLOW_METHOD(default_allow_method, std::end(default_allow_method));
 const std::string ConfigConstant::DEFAULT_AUTOINDEX = "off";
-const int ConfigConstant::DEFAULT_CLIENT_MAX_BODY_SIZE = 5000;
+const size_t ConfigConstant::DEFAULT_CLIENT_MAX_BODY_SIZE = 5000;
 const std::vector<std::string>
     ConfigConstant::DEFAULT_INDEX(default_index, std::end(default_index));
-const int ConfigConstant::DEFAULT_PORT = 80;
+const size_t ConfigConstant::DEFAULT_PORT = 80;
